WIP: cleanup and docks, replace old tool with new
