+++ simulator_tests/__init__.py
@@ -16,7 +16,6 @@ from .test_cross_tool_comprehensive import CrossToolComprehensiveTest
 from .test_cross_tool_continuation import CrossToolContinuationTest
 from .test_debug_certain_confidence import DebugCertainConfidenceTest
 from .test_debug_validation import DebugValidationTest
-from .test_debugworkflow_validation import DebugWorkflowValidationTest
 from .test_line_number_validation import LineNumberValidationTest
 from .test_logs_validation import LogsValidationTest
 from .test_model_thinking_config import TestModelThinkingConfig
@@ -57,7 +56,6 @@ TEST_REGISTRY = {
     "testgen_validation": TestGenValidationTest,
     "refactor_validation": RefactorValidationTest,
     "debug_validation": DebugValidationTest,
-    "debugworkflow_validation": DebugWorkflowValidationTest,
     "debug_certain_confidence": DebugCertainConfidenceTest,
     "conversation_chain_validation": ConversationChainValidationTest,
     "vision_capability": VisionCapabilityTest,
@@ -90,7 +88,6 @@ __all__ = [
     "TestGenValidationTest",
     "RefactorValidationTest",
     "DebugValidationTest",
-    "DebugWorkflowValidationTest",
     "DebugCertainConfidenceTest",
     "ConversationChainValidationTest",
     "VisionCapabilityTest",
