+++ tools/debug.py
@@ -1,42 +1,58 @@
 """
-Debug Issue tool - Root cause analysis and debugging assistance with systematic investigation
+Debug tool - Systematic root cause analysis and debugging assistance
+
+This tool provides a structured workflow for investigating complex bugs and issues.
+It guides you through systematic investigation steps with forced pauses between each step
+to ensure thorough code examination before proceeding. The tool supports backtracking,
+hypothesis evolution, and expert analysis integration for comprehensive debugging.
+
+Key features:
+- Step-by-step investigation workflow with progress tracking
+- Context-aware file embedding (references during investigation, full content for analysis)
+- Automatic conversation threading and history preservation
+- Expert analysis integration with external models
+- Support for visual debugging with image context
+- Confidence-based workflow optimization
 """
 
-import json
 import logging
 from typing import TYPE_CHECKING, Any, Optional
 
-from pydantic import Field, field_validator
+from pydantic import Field, field_validator, model_validator
 
 if TYPE_CHECKING:
     from tools.models import ToolModelCategory
 
 from config import TEMPERATURE_ANALYTICAL
 from systemprompts import DEBUG_ISSUE_PROMPT
+from tools.shared.base_models import WorkflowRequest
 
-from .base import BaseTool, ToolRequest
+from .workflow.base import WorkflowTool
 
 logger = logging.getLogger(__name__)
 
-# Field descriptions for the investigation steps
+# Tool-specific field descriptions matching original debug tool
 DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS = {
     "step": (
         "Describe what you're currently investigating by thinking deeply about the issue and its possible causes. "
-        "In step 1, clearly state the issue and begin forming an investigative direction. CRITICAL: Remember that "
-        "reported symptoms might originate from code far from where they manifest. Also be aware that after thorough "
-        "investigation, you might find NO BUG EXISTS - it could be a misunderstanding or expectation mismatch. "
-        "Consider not only obvious failures, but also subtle contributing factors like upstream logic, invalid inputs, "
-        "missing preconditions, or hidden side effects. Map out the flow of related functions or modules. Identify "
-        "call paths where input values or branching logic could cause instability. In concurrent systems, watch for "
-        "race conditions, shared state, or timing dependencies. In all later steps, continue exploring with precision: "
-        "trace deeper dependencies, verify hypotheses, and adapt your understanding as you uncover more evidence."
+        "In step 1, clearly state the issue and begin forming an investigative direction after thinking carefully"
+        "about the described problem. Ask further questions from the user if you think these will help with your"
+        "understanding and investigation. CRITICAL: Remember that reported symptoms might originate from code far from "
+        "where they manifest. Also be aware that after thorough investigation, you might find NO BUG EXISTS - it could "
+        "be a misunderstanding or expectation mismatch. Consider not only obvious failures, but also subtle "
+        "contributing factors like upstream logic, invalid inputs, missing preconditions, or hidden side effects. "
+        "Map out the flow of related functions or modules. Identify call paths where input values or branching logic "
+        "could cause instability. In concurrent systems, watch for race conditions, shared state, or timing "
+        "dependencies. In all later steps, continue exploring with precision: trace deeper dependencies, verify "
+        "hypotheses, and adapt your understanding as you uncover more evidence."
     ),
     "step_number": (
         "The index of the current step in the investigation sequence, beginning at 1. Each step should build upon or "
         "revise the previous one."
     ),
     "total_steps": (
-        "Your current estimate for how many steps will be needed to complete the investigation. Adjust as new findings emerge."
+        "Your current estimate for how many steps will be needed to complete the investigation. "
+        "Adjust as new findings emerge."
     ),
     "next_step_required": (
         "Set to true if you plan to continue the investigation with another step. False means you believe the root "
@@ -46,11 +62,13 @@ DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS = {
         "Summarize everything discovered in this step. Include new clues, unexpected behavior, evidence from code or "
         "logs, or disproven theories. Be specific and avoid vague language—document what you now know and how it "
         "affects your hypothesis. IMPORTANT: If you find no evidence supporting the reported issue after thorough "
-        "investigation, document this clearly. Finding 'no bug' is a valid outcome if the investigation was comprehensive. "
+        "investigation, document this clearly. Finding 'no bug' is a valid outcome if the "
+        "investigation was comprehensive. "
         "In later steps, confirm or disprove past findings with reason."
     ),
     "files_checked": (
-        "List all files (as absolute paths, do not clip or shrink file names) examined during the investigation so far. "
+        "List all files (as absolute paths, do not clip or shrink file names) examined during "
+        "the investigation so far. "
         "Include even files ruled out, as this tracks your exploration path."
     ),
     "relevant_files": (
@@ -58,8 +76,9 @@ DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS = {
         "those that are directly tied to the root cause or its effects. This could include the cause, trigger, or "
         "place of manifestation."
     ),
-    "relevant_methods": (
-        "List methods or functions that are central to the issue, in the format 'ClassName.methodName' or 'functionName'. "
+    "relevant_context": (
+        "List methods or functions that are central to the issue, in the format "
+        "'ClassName.methodName' or 'functionName'. "
         "Prioritize those that influence or process inputs, drive branching, or pass state between modules."
     ),
     "hypothesis": (
@@ -72,7 +91,8 @@ DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS = {
     ),
     "confidence": (
         "Indicate your current confidence in the hypothesis. Use: 'exploring' (starting out), 'low' (early idea), "
-        "'medium' (some supporting evidence), 'high' (strong evidence), 'certain' (only when the root cause and minimal "
+        "'medium' (some supporting evidence), 'high' (strong evidence), 'certain' (only when "
+        "the root cause and minimal "
         "fix are both confirmed). Do NOT use 'certain' unless the issue can be fully resolved with a fix, use 'high' "
         "instead when in doubt. Using 'certain' prevents you from taking assistance from another thought-partner."
     ),
@@ -80,29 +100,15 @@ DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS = {
         "If an earlier finding or hypothesis needs to be revised or discarded, specify the step number from which to "
         "start over. Use this to acknowledge investigative dead ends and correct the course."
     ),
-    "continuation_id": "Continuation token used for linking multi-step investigations and continuing conversations after discovery.",
     "images": (
         "Optional list of absolute paths to screenshots or UI visuals that clarify the issue. "
         "Only include if they materially assist understanding or hypothesis formulation."
     ),
 }
 
-DEBUG_FIELD_DESCRIPTIONS = {
-    "initial_issue": "Describe the original problem that triggered the investigation.",
-    "investigation_summary": (
-        "Full overview of the systematic investigation process. Reflect deep thinking and each step's contribution to narrowing down the issue."
-    ),
-    "findings": "Final list of critical insights and discoveries across all steps.",
-    "files": "Essential files referenced during investigation (must be full absolute paths).",
-    "error_context": "Logs, tracebacks, or execution details that support the root cause hypothesis.",
-    "relevant_methods": "List of all methods/functions identified as directly involved.",
-    "hypothesis": "Final, most likely explanation of the root cause based on evidence.",
-    "images": "Optional screenshots or visual materials that helped diagnose the issue.",
-}
 
-
-class DebugInvestigationRequest(ToolRequest):
-    """Request model for debug investigation steps"""
+class DebugInvestigationRequest(WorkflowRequest):
+    """Request model for debug investigation steps matching original debug tool exactly"""
 
     # Required fields for each investigation step
     step: str = Field(..., description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["step"])
@@ -118,8 +124,11 @@ class DebugInvestigationRequest(ToolRequest):
     relevant_files: list[str] = Field(
         default_factory=list, description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["relevant_files"]
     )
+    relevant_context: list[str] = Field(
+        default_factory=list, description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["relevant_context"]
+    )
     relevant_methods: list[str] = Field(
-        default_factory=list, description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["relevant_methods"]
+        default_factory=list, description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["relevant_context"], exclude=True
     )
     hypothesis: Optional[str] = Field(None, description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["hypothesis"])
     confidence: Optional[str] = Field("low", description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["confidence"])
@@ -129,9 +138,6 @@ class DebugInvestigationRequest(ToolRequest):
         None, description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["backtrack_from_step"]
     )
 
-    # Optional continuation field
-    continuation_id: Optional[str] = Field(None, description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["continuation_id"])
-
     # Optional images for visual debugging
     images: Optional[list[str]] = Field(default=None, description=DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["images"])
 
@@ -140,30 +146,30 @@ class DebugInvestigationRequest(ToolRequest):
     thinking_mode: Optional[str] = Field(default=None, exclude=True)
     use_websearch: Optional[bool] = Field(default=None, exclude=True)
 
-    @field_validator("files_checked", "relevant_files", "relevant_methods", mode="before")
-    @classmethod
-    def convert_string_to_list(cls, v):
-        """Convert string inputs to empty lists to handle malformed inputs gracefully."""
-        if isinstance(v, str):
-            logger.warning(f"Field received string '{v}' instead of list, converting to empty list")
-            return []
-        return v
+    @model_validator(mode="after")
+    def map_relevant_methods_to_context(self):
+        """Map relevant_methods from external input to relevant_context for internal processing."""
+        # If relevant_context is empty but relevant_methods has values, use relevant_methods
+        if not self.relevant_context and self.relevant_methods:
+            self.relevant_context = self.relevant_methods[:]
+        return self
 
 
-class DebugIssueTool(BaseTool):
-    """Advanced debugging tool with systematic self-investigation"""
+
+
+class DebugIssueTool(WorkflowTool):
+    """
+    Debug tool for systematic root cause analysis and issue investigation.
+
+    This tool implements a structured debugging workflow that guides users through
+    methodical investigation steps, ensuring thorough code examination and evidence
+    gathering before reaching conclusions. It supports complex debugging scenarios
+    including race conditions, memory leaks, performance issues, and integration problems.
+    """
 
     def __init__(self):
         super().__init__()
-        self.investigation_history = []
-        self.consolidated_findings = {
-            "files_checked": set(),
-            "relevant_files": set(),
-            "relevant_methods": set(),
-            "findings": [],
-            "hypotheses": [],
-            "images": [],
-        }
+        self.initial_issue = None
 
     def get_name(self) -> str:
         return "debug"
@@ -189,79 +195,6 @@ class DebugIssueTool(BaseTool):
             "race conditions, memory leaks, integration problems."
         )
 
-    def get_input_schema(self) -> dict[str, Any]:
-        schema = {
-            "type": "object",
-            "properties": {
-                # Investigation step fields
-                "step": {
-                    "type": "string",
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["step"],
-                },
-                "step_number": {
-                    "type": "integer",
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["step_number"],
-                    "minimum": 1,
-                },
-                "total_steps": {
-                    "type": "integer",
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["total_steps"],
-                    "minimum": 1,
-                },
-                "next_step_required": {
-                    "type": "boolean",
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["next_step_required"],
-                },
-                "findings": {
-                    "type": "string",
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["findings"],
-                },
-                "files_checked": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["files_checked"],
-                },
-                "relevant_files": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["relevant_files"],
-                },
-                "relevant_methods": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["relevant_methods"],
-                },
-                "hypothesis": {
-                    "type": "string",
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["hypothesis"],
-                },
-                "confidence": {
-                    "type": "string",
-                    "enum": ["exploring", "low", "medium", "high", "certain"],
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["confidence"],
-                },
-                "backtrack_from_step": {
-                    "type": "integer",
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["backtrack_from_step"],
-                    "minimum": 1,
-                },
-                "continuation_id": {
-                    "type": "string",
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["continuation_id"],
-                },
-                "images": {
-                    "type": "array",
-                    "items": {"type": "string"},
-                    "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["images"],
-                },
-                # Add model field for proper model selection
-                "model": self.get_model_field_schema(),
-            },
-            # Required fields for investigation
-            "required": ["step", "step_number", "total_steps", "next_step_required", "findings"]
-            + (["model"] if self.is_effective_auto_mode() else []),
-        }
-        return schema
 
     def get_system_prompt(self) -> str:
         return DEBUG_ISSUE_PROMPT
@@ -272,462 +205,407 @@ class DebugIssueTool(BaseTool):
     def get_model_category(self) -> "ToolModelCategory":
         """Debug requires deep analysis and reasoning"""
         from tools.models import ToolModelCategory
-
         return ToolModelCategory.EXTENDED_REASONING
 
-    def get_request_model(self):
+    def get_workflow_request_model(self):
+        """Return the debug-specific request model."""
         return DebugInvestigationRequest
 
-    def requires_model(self) -> bool:
-        """
-        Debug tool requires a model for expert analysis after investigation.
-        """
-        return True
+    def get_input_schema(self) -> dict[str, Any]:
+        """Generate input schema using WorkflowSchemaBuilder with debug-specific overrides."""
+        from .workflow.schema_builders import WorkflowSchemaBuilder
 
-    async def execute(self, arguments: dict[str, Any]) -> list:
-        """
-        Override execute to implement self-investigation pattern.
-
-        Investigation Flow:
-        1. Claude calls debug with investigation steps
-        2. Tool tracks findings, files, methods progressively
-        3. Once investigation is complete, tool calls AI model for expert analysis
-        4. Returns structured response combining investigation + expert analysis
-        """
-        from mcp.types import TextContent
-
-        from utils.conversation_memory import add_turn, create_thread
-
-        try:
-            # Validate request
-            request = DebugInvestigationRequest(**arguments)
-
-            # Adjust total steps if needed
-            if request.step_number > request.total_steps:
-                request.total_steps = request.step_number
-
-            # Handle continuation
-            continuation_id = request.continuation_id
-
-            # Create thread for first step
-            if not continuation_id and request.step_number == 1:
-                # Clean arguments to remove non-serializable fields
-                clean_args = {k: v for k, v in arguments.items() if k not in ["_model_context", "_resolved_model_name"]}
-                continuation_id = create_thread("debug", clean_args)
-                # Store initial issue description
-                self.initial_issue = request.step
-
-            # Handle backtracking first if requested
-            if request.backtrack_from_step:
-                # Remove findings after the backtrack point
-                self.investigation_history = [
-                    s for s in self.investigation_history if s["step_number"] < request.backtrack_from_step
-                ]
-                # Reprocess consolidated findings to match truncated history
-                self._reprocess_consolidated_findings()
-
-                # Log if step number needs correction
-                expected_step_number = len(self.investigation_history) + 1
-                if request.step_number != expected_step_number:
-                    logger.debug(
-                        f"Step number adjusted from {request.step_number} to {expected_step_number} after backtracking"
-                    )
-
-            # Process investigation step
-            step_data = {
-                "step": request.step,
-                "step_number": request.step_number,
-                "findings": request.findings,
-                "files_checked": request.files_checked,
-                "relevant_files": request.relevant_files,
-                "relevant_methods": request.relevant_methods,
-                "hypothesis": request.hypothesis,
-                "confidence": request.confidence,
-                "images": request.images,
-            }
-
-            # Store in history
-            self.investigation_history.append(step_data)
-
-            # Update consolidated findings
-            self.consolidated_findings["files_checked"].update(request.files_checked)
-            self.consolidated_findings["relevant_files"].update(request.relevant_files)
-            self.consolidated_findings["relevant_methods"].update(request.relevant_methods)
-            self.consolidated_findings["findings"].append(f"Step {request.step_number}: {request.findings}")
-            if request.hypothesis:
-                self.consolidated_findings["hypotheses"].append(
-                    {"step": request.step_number, "hypothesis": request.hypothesis, "confidence": request.confidence}
-                )
-            if request.images:
-                self.consolidated_findings["images"].extend(request.images)
-
-            # Build response
-            response_data = {
-                "status": "investigation_in_progress",
-                "step_number": request.step_number,
-                "total_steps": request.total_steps,
-                "next_step_required": request.next_step_required,
-                "investigation_status": {
-                    "files_checked": len(self.consolidated_findings["files_checked"]),
-                    "relevant_files": len(self.consolidated_findings["relevant_files"]),
-                    "relevant_methods": len(self.consolidated_findings["relevant_methods"]),
-                    "hypotheses_formed": len(self.consolidated_findings["hypotheses"]),
-                    "images_collected": len(set(self.consolidated_findings["images"])),
-                    "current_confidence": request.confidence,
-                },
-            }
-
-            if continuation_id:
-                response_data["continuation_id"] = continuation_id
-
-            # If investigation is complete, decide whether to call expert analysis or proceed with minimal fix
-            if not request.next_step_required:
-                response_data["investigation_complete"] = True
-
-                # Check if Claude has absolute certainty and can proceed with minimal fix
-                if request.confidence == "certain":
-                    # Trust Claude's judgment completely - if it says certain, skip expert analysis
-                    response_data["status"] = "certain_confidence_proceed_with_fix"
-
-                    investigation_summary = self._prepare_investigation_summary()
-                    response_data["complete_investigation"] = {
-                        "initial_issue": getattr(self, "initial_issue", request.step),
-                        "steps_taken": len(self.investigation_history),
-                        "files_examined": list(self.consolidated_findings["files_checked"]),
-                        "relevant_files": list(self.consolidated_findings["relevant_files"]),
-                        "relevant_methods": list(self.consolidated_findings["relevant_methods"]),
-                        "investigation_summary": investigation_summary,
-                        "final_hypothesis": request.hypothesis,
-                        "confidence_level": "certain",
-                    }
-                    response_data["next_steps"] = (
-                        "Investigation complete with CERTAIN confidence. You have identified the exact "
-                        "root cause and a minimal fix. MANDATORY: Present the user with the root cause analysis"
-                        "and IMMEDIATELY proceed with implementing the simple fix without requiring further "
-                        "consultation. Focus on the precise, minimal change needed."
-                    )
-                    response_data["skip_expert_analysis"] = True
-                    response_data["expert_analysis"] = {
-                        "status": "skipped_due_to_certain_confidence",
-                        "reason": "Claude identified exact root cause with minimal fix requirement",
-                    }
-                else:
-                    # Standard expert analysis for certain/high/medium/low/exploring confidence
-                    response_data["status"] = "calling_expert_analysis"
-
-                    # Prepare consolidated investigation summary
-                    investigation_summary = self._prepare_investigation_summary()
-
-                    # Call the AI model with full context
-                    expert_analysis = await self._call_expert_analysis(
-                        initial_issue=getattr(self, "initial_issue", request.step),
-                        investigation_summary=investigation_summary,
-                        relevant_files=list(self.consolidated_findings["relevant_files"]),
-                        relevant_methods=list(self.consolidated_findings["relevant_methods"]),
-                        final_hypothesis=request.hypothesis,
-                        error_context=self._extract_error_context(),
-                        images=list(set(self.consolidated_findings["images"])),  # Unique images
-                        model_info=arguments.get("_model_context"),  # Use pre-resolved model context from server.py
-                        arguments=arguments,  # Pass arguments for model resolution
-                        request=request,  # Pass request for model resolution
-                    )
-
-                    # Combine investigation and expert analysis
-                    response_data["expert_analysis"] = expert_analysis
-                    response_data["complete_investigation"] = {
-                        "initial_issue": getattr(self, "initial_issue", request.step),
-                        "steps_taken": len(self.investigation_history),
-                        "files_examined": list(self.consolidated_findings["files_checked"]),
-                        "relevant_files": list(self.consolidated_findings["relevant_files"]),
-                        "relevant_methods": list(self.consolidated_findings["relevant_methods"]),
-                        "investigation_summary": investigation_summary,
-                    }
-                    response_data["next_steps"] = (
-                        "INVESTIGATION IS COMPLETE. YOU MUST now summarize and present ALL key findings, confirmed "
-                        "hypotheses, and exact recommended fixes. Clearly identify the most likely root cause and "
-                        "provide concrete, actionable implementation guidance. Highlight affected code paths and display "
-                        "reasoning that led to this conclusion—make it easy for a developer to understand exactly where "
-                        "the problem lies."
-                    )
-            else:
-                # CRITICAL: Force Claude to actually investigate before calling debug again
-                response_data["status"] = "pause_for_investigation"
-                response_data["investigation_required"] = True
-
-                if request.step_number == 1:
-                    # Initial investigation tasks
-                    response_data["required_actions"] = [
-                        "Search for code related to the reported issue or symptoms",
-                        "Examine relevant files and understand the current implementation",
-                        "Understand the project structure and locate relevant modules",
-                        "Identify how the affected functionality is supposed to work",
-                    ]
-                    response_data["next_steps"] = (
-                        f"MANDATORY: DO NOT call the debug tool again immediately. You MUST first investigate "
-                        f"the codebase using appropriate tools. CRITICAL AWARENESS: The reported symptoms might be "
-                        f"caused by issues elsewhere in the code, not where symptoms appear. Also, after thorough "
-                        f"investigation, it's possible NO BUG EXISTS - the issue might be a misunderstanding or "
-                        f"user expectation mismatch. Search broadly, examine implementations, understand the logic flow. "
-                        f"Only call debug again AFTER gathering concrete evidence. When you call debug next time, "
-                        f"use step_number: {request.step_number + 1} and report specific files examined and findings discovered."
-                    )
-                elif request.step_number >= 2 and request.confidence in ["exploring", "low"]:
-                    # Need deeper investigation
-                    response_data["required_actions"] = [
-                        "Examine the specific files you've identified as relevant",
-                        "Trace method calls and data flow through the system",
-                        "Check for edge cases, boundary conditions, and assumptions in the code",
-                        "Look for related configuration, dependencies, or external factors",
-                    ]
-                    response_data["next_steps"] = (
-                        f"STOP! Do NOT call debug again yet. Based on your findings, you've identified potential areas "
-                        f"but need concrete evidence. MANDATORY ACTIONS before calling debug step {request.step_number + 1}:\n"
-                        f"1. Examine ALL files in your relevant_files list\n"
-                        f"2. Trace how data flows through {', '.join(request.relevant_methods[:3]) if request.relevant_methods else 'the identified components'}\n"
-                        f"3. Look for logic errors, incorrect assumptions, missing validations\n"
-                        f"4. Check interactions between components and external dependencies\n"
-                        f"Only call debug again with step_number: {request.step_number + 1} AFTER completing these investigations."
-                    )
-                elif request.confidence in ["medium", "high"]:
-                    # Close to root cause - need confirmation
-                    response_data["required_actions"] = [
-                        "Examine the exact code sections where you believe the issue occurs",
-                        "Trace the execution path that leads to the failure",
-                        "Verify your hypothesis with concrete code evidence",
-                        "Check for any similar patterns elsewhere in the codebase",
-                    ]
-                    response_data["next_steps"] = (
-                        f"WAIT! Your hypothesis needs verification. DO NOT call debug immediately. REQUIRED ACTIONS:\n"
-                        f"1. Examine the exact lines where the issue occurs\n"
-                        f"2. Trace backwards: how does data get to this point? What transforms it?\n"
-                        f"3. Check all assumptions: are inputs validated? Are nulls handled?\n"
-                        f"4. Look for the EXACT line where expected != actual behavior\n"
-                        f"REMEMBER: If you cannot find concrete evidence of a bug causing the reported symptoms, "
-                        f"'no bug found' is a valid conclusion. Consider suggesting discussion with your thought partner "
-                        f"or engineering assistant for clarification. Document findings with specific file:line references, "
-                        f"then call debug with step_number: {request.step_number + 1}."
-                    )
-                else:
-                    # General investigation needed
-                    response_data["required_actions"] = [
-                        "Continue examining the code paths identified in your hypothesis",
-                        "Gather more evidence using appropriate investigation tools",
-                        "Test edge cases and boundary conditions",
-                        "Look for patterns that confirm or refute your theory",
-                    ]
-                    response_data["next_steps"] = (
-                        f"PAUSE INVESTIGATION. Before calling debug step {request.step_number + 1}, you MUST examine code. "
-                        f"Required: Read files from your files_checked list, search for patterns in your hypothesis, "
-                        f"trace execution flow. Your next debug call (step_number: {request.step_number + 1}) must include "
-                        f"NEW evidence from actual code examination, not just theories. If no bug evidence is found, suggesting "
-                        f"collaboration with thought partner is valuable. NO recursive debug calls without investigation work!"
-                    )
-
-            # Store in conversation memory
-            if continuation_id:
-                add_turn(
-                    thread_id=continuation_id,
-                    role="assistant",
-                    content=json.dumps(response_data, indent=2),
-                    tool_name="debug",
-                    files=list(self.consolidated_findings["relevant_files"]),
-                    images=request.images,
-                )
-
-            return [TextContent(type="text", text=json.dumps(response_data, indent=2))]
-
-        except Exception as e:
-            logger.error(f"Error in debug investigation: {e}", exc_info=True)
-            error_data = {
-                "status": "investigation_failed",
-                "error": str(e),
-                "step_number": arguments.get("step_number", 0),
-            }
-            return [TextContent(type="text", text=json.dumps(error_data, indent=2))]
-
-    def _reprocess_consolidated_findings(self):
-        """Reprocess consolidated findings after backtracking"""
-        self.consolidated_findings = {
-            "files_checked": set(),
-            "relevant_files": set(),
-            "relevant_methods": set(),
-            "findings": [],
-            "hypotheses": [],
-            "images": [],
+        # Debug-specific field overrides
+        debug_field_overrides = {
+            "step": {
+                "type": "string",
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["step"],
+            },
+            "step_number": {
+                "type": "integer",
+                "minimum": 1,
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["step_number"],
+            },
+            "total_steps": {
+                "type": "integer",
+                "minimum": 1,
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["total_steps"],
+            },
+            "next_step_required": {
+                "type": "boolean",
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["next_step_required"],
+            },
+            "findings": {
+                "type": "string",
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["findings"],
+            },
+            "files_checked": {
+                "type": "array",
+                "items": {"type": "string"},
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["files_checked"],
+            },
+            "relevant_files": {
+                "type": "array",
+                "items": {"type": "string"},
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["relevant_files"],
+            },
+            "confidence": {
+                "type": "string",
+                "enum": ["exploring", "low", "medium", "high", "certain"],
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["confidence"],
+            },
+            "hypothesis": {
+                "type": "string",
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["hypothesis"],
+            },
+            "backtrack_from_step": {
+                "type": "integer",
+                "minimum": 1,
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["backtrack_from_step"],
+            },
+            "relevant_methods": {
+                "type": "array",
+                "items": {"type": "string"},
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["relevant_context"],
+            },
+            "images": {
+                "type": "array",
+                "items": {"type": "string"},
+                "description": DEBUG_INVESTIGATION_FIELD_DESCRIPTIONS["images"],
+            },
         }
 
-        for step in self.investigation_history:
-            self.consolidated_findings["files_checked"].update(step.get("files_checked", []))
-            self.consolidated_findings["relevant_files"].update(step.get("relevant_files", []))
-            self.consolidated_findings["relevant_methods"].update(step.get("relevant_methods", []))
-            self.consolidated_findings["findings"].append(f"Step {step['step_number']}: {step['findings']}")
-            if step.get("hypothesis"):
-                self.consolidated_findings["hypotheses"].append(
-                    {
-                        "step": step["step_number"],
-                        "hypothesis": step["hypothesis"],
-                        "confidence": step.get("confidence", "low"),
-                    }
-                )
-            if step.get("images"):
-                self.consolidated_findings["images"].extend(step["images"])
+        # Use WorkflowSchemaBuilder with debug-specific tool fields
+        return WorkflowSchemaBuilder.build_schema(
+            tool_specific_fields=debug_field_overrides,
+            model_field_schema=self.get_model_field_schema(),
+            auto_mode=self.is_effective_auto_mode(),
+            tool_name=self.get_name(),
+        )
 
-    def _prepare_investigation_summary(self) -> str:
-        """Prepare a comprehensive summary of the investigation"""
+    def get_required_actions(self, step_number: int, confidence: str, findings: str, total_steps: int) -> list[str]:
+        """Define required actions for each investigation phase."""
+        if step_number == 1:
+            # Initial investigation tasks
+            return [
+                "Search for code related to the reported issue or symptoms",
+                "Examine relevant files and understand the current implementation",
+                "Understand the project structure and locate relevant modules",
+                "Identify how the affected functionality is supposed to work",
+            ]
+        elif confidence in ["exploring", "low"]:
+            # Need deeper investigation
+            return [
+                "Examine the specific files you've identified as relevant",
+                "Trace method calls and data flow through the system",
+                "Check for edge cases, boundary conditions, and assumptions in the code",
+                "Look for related configuration, dependencies, or external factors",
+            ]
+        elif confidence in ["medium", "high"]:
+            # Close to root cause - need confirmation
+            return [
+                "Examine the exact code sections where you believe the issue occurs",
+                "Trace the execution path that leads to the failure",
+                "Verify your hypothesis with concrete code evidence",
+                "Check for any similar patterns elsewhere in the codebase",
+            ]
+        else:
+            # General investigation needed
+            return [
+                "Continue examining the code paths identified in your hypothesis",
+                "Gather more evidence using appropriate investigation tools",
+                "Test edge cases and boundary conditions",
+                "Look for patterns that confirm or refute your theory",
+            ]
+
+    def should_call_expert_analysis(self, consolidated_findings) -> bool:
+        """
+        Decide when to call external model based on investigation completeness.
+
+        Don't call expert analysis if Claude has certain confidence - trust their judgment.
+        """
+        # Check if we have meaningful investigation data
+        return (
+            len(consolidated_findings.relevant_files) > 0
+            or len(consolidated_findings.findings) >= 2
+            or len(consolidated_findings.issues_found) > 0
+        )
+
+    def prepare_expert_analysis_context(self, consolidated_findings) -> str:
+        """Prepare context for external model call matching original debug tool format."""
+        context_parts = [
+            f"=== ISSUE DESCRIPTION ===\n{self.initial_issue or 'Investigation initiated'}\n=== END DESCRIPTION ==="
+        ]
+
+        # Add investigation summary
+        investigation_summary = self._build_investigation_summary(consolidated_findings)
+        context_parts.append(
+            f"\n=== CLAUDE'S INVESTIGATION FINDINGS ===\n{investigation_summary}\n=== END FINDINGS ==="
+        )
+
+        # Add error context if available
+        error_context = self._extract_error_context(consolidated_findings)
+        if error_context:
+            context_parts.append(f"\n=== ERROR CONTEXT/STACK TRACE ===\n{error_context}\n=== END CONTEXT ===")
+
+        # Add relevant methods if available (map relevant_context back to relevant_methods)
+        if consolidated_findings.relevant_context:
+            methods_text = "\n".join(f"- {method}" for method in consolidated_findings.relevant_context)
+            context_parts.append(f"\n=== RELEVANT METHODS/FUNCTIONS ===\n{methods_text}\n=== END METHODS ===")
+
+        # Add hypothesis evolution if available
+        if consolidated_findings.hypotheses:
+            hypotheses_text = "\n".join(
+                f"Step {h['step']} ({h['confidence']} confidence): {h['hypothesis']}"
+                for h in consolidated_findings.hypotheses
+            )
+            context_parts.append(f"\n=== HYPOTHESIS EVOLUTION ===\n{hypotheses_text}\n=== END HYPOTHESES ===")
+
+        # Add images if available
+        if consolidated_findings.images:
+            images_text = "\n".join(f"- {img}" for img in consolidated_findings.images)
+            context_parts.append(
+                f"\n=== VISUAL DEBUGGING INFORMATION ===\n{images_text}\n=== END VISUAL INFORMATION ==="
+            )
+
+        # Add file content if we have relevant files
+        if consolidated_findings.relevant_files:
+            file_content, _ = self._prepare_file_content_for_prompt(
+                list(consolidated_findings.relevant_files),
+                None,
+                "Essential debugging files"
+            )
+            if file_content:
+                context_parts.append(
+                    f"\n=== ESSENTIAL FILES FOR DEBUGGING ===\n{file_content}\n=== END ESSENTIAL FILES ==="
+                )
+
+        return "\n".join(context_parts)
+
+    def _build_investigation_summary(self, consolidated_findings) -> str:
+        """Prepare a comprehensive summary of the investigation."""
         summary_parts = [
             "=== SYSTEMATIC INVESTIGATION SUMMARY ===",
-            f"Total steps: {len(self.investigation_history)}",
-            f"Files examined: {len(self.consolidated_findings['files_checked'])}",
-            f"Relevant files identified: {len(self.consolidated_findings['relevant_files'])}",
-            f"Methods/functions involved: {len(self.consolidated_findings['relevant_methods'])}",
+            f"Total steps: {len(consolidated_findings.findings)}",
+            f"Files examined: {len(consolidated_findings.files_checked)}",
+            f"Relevant files identified: {len(consolidated_findings.relevant_files)}",
+            f"Methods/functions involved: {len(consolidated_findings.relevant_context)}",
             "",
             "=== INVESTIGATION PROGRESSION ===",
         ]
 
-        for finding in self.consolidated_findings["findings"]:
+        for finding in consolidated_findings.findings:
             summary_parts.append(finding)
-
-        if self.consolidated_findings["hypotheses"]:
-            summary_parts.extend(
-                [
-                    "",
-                    "=== HYPOTHESIS EVOLUTION ===",
-                ]
-            )
-            for hyp in self.consolidated_findings["hypotheses"]:
-                summary_parts.append(f"Step {hyp['step']} ({hyp['confidence']} confidence): {hyp['hypothesis']}")
 
         return "\n".join(summary_parts)
 
-    def _extract_error_context(self) -> Optional[str]:
-        """Extract error context from investigation findings"""
+    def _extract_error_context(self, consolidated_findings) -> Optional[str]:
+        """Extract error context from investigation findings."""
         error_patterns = ["error", "exception", "stack trace", "traceback", "failure"]
         error_context_parts = []
 
-        for finding in self.consolidated_findings["findings"]:
+        for finding in consolidated_findings.findings:
             if any(pattern in finding.lower() for pattern in error_patterns):
                 error_context_parts.append(finding)
 
         return "\n".join(error_context_parts) if error_context_parts else None
 
-    async def _call_expert_analysis(
-        self,
-        initial_issue: str,
-        investigation_summary: str,
-        relevant_files: list[str],
-        relevant_methods: list[str],
-        final_hypothesis: Optional[str],
-        error_context: Optional[str],
-        images: list[str],
-        model_info: Optional[Any] = None,
-        arguments: Optional[dict] = None,
-        request: Optional[Any] = None,
-    ) -> dict:
-        """Call AI model for expert analysis of the investigation"""
-        # Set up model context when we actually need it for expert analysis
-        # Use the same model resolution logic as the base class
-        if model_info:
-            # Use pre-resolved model context from server.py (normal case)
-            self._model_context = model_info
-            model_name = model_info.model_name
+    def get_step_guidance(self, step_number: int, confidence: str, request) -> dict[str, Any]:
+        """
+        Provide step-specific guidance matching original debug tool behavior.
+
+        This method generates debug-specific guidance that's used by get_step_guidance_message().
+        """
+        # Generate the next steps instruction based on required actions
+        required_actions = self.get_required_actions(step_number, confidence, request.findings, request.total_steps)
+
+        if step_number == 1:
+            next_steps = (
+                f"MANDATORY: DO NOT call the {self.get_name()} tool again immediately. You MUST first investigate "
+                f"the codebase using appropriate tools. CRITICAL AWARENESS: The reported symptoms might be "
+                f"caused by issues elsewhere in the code, not where symptoms appear. Also, after thorough "
+                f"investigation, it's possible NO BUG EXISTS - the issue might be a misunderstanding or "
+                f"user expectation mismatch. Search broadly, examine implementations, understand the logic flow. "
+                f"Only call {self.get_name()} again AFTER gathering concrete evidence. When you call "
+                f"{self.get_name()} next time, "
+                f"use step_number: {step_number + 1} and report specific files examined and findings discovered."
+            )
+        elif confidence in ["exploring", "low"]:
+            next_steps = (
+                f"STOP! Do NOT call {self.get_name()} again yet. Based on your findings, you've identified potential areas "
+                f"but need concrete evidence. MANDATORY ACTIONS before calling {self.get_name()} step {step_number + 1}:\n"
+                + "\n".join(f"{i+1}. {action}" for i, action in enumerate(required_actions))
+                + f"\n\nOnly call {self.get_name()} again with step_number: {step_number + 1} AFTER "
+                + "completing these investigations."
+            )
+        elif confidence in ["medium", "high"]:
+            next_steps = (
+                f"WAIT! Your hypothesis needs verification. DO NOT call {self.get_name()} immediately. REQUIRED ACTIONS:\n"
+                + "\n".join(f"{i+1}. {action}" for i, action in enumerate(required_actions))
+                + f"\n\nREMEMBER: If you cannot find concrete evidence of a bug causing the reported symptoms, "
+                f"'no bug found' is a valid conclusion. Consider suggesting discussion with your thought partner "
+                f"or engineering assistant for clarification. Document findings with specific file:line references, "
+                f"then call {self.get_name()} with step_number: {step_number + 1}."
+            )
         else:
-            # Use centralized model resolution from base class
-            if arguments and request:
-                try:
-                    model_name, model_context = self._resolve_model_context(arguments, request)
-                    self._model_context = model_context
-                except ValueError as e:
-                    # Model resolution failed, return error
-                    return {"error": f"Model resolution failed: {str(e)}", "status": "model_resolution_error"}
-            else:
-                # Last resort fallback if no arguments/request provided
-                from config import DEFAULT_MODEL
-                from utils.model_context import ModelContext
-
-                model_name = DEFAULT_MODEL
-                self._model_context = ModelContext(model_name)
-
-        # Store model name for use by other methods
-        self._current_model_name = model_name
-        provider = self.get_model_provider(model_name)
-
-        # Prepare the debug prompt with all investigation context
-        prompt_parts = [
-            f"=== ISSUE DESCRIPTION ===\n{initial_issue}\n=== END DESCRIPTION ===",
-            f"\n=== CLAUDE'S INVESTIGATION FINDINGS ===\n{investigation_summary}\n=== END FINDINGS ===",
-        ]
-
-        if error_context:
-            prompt_parts.append(f"\n=== ERROR CONTEXT/STACK TRACE ===\n{error_context}\n=== END CONTEXT ===")
-
-        if relevant_methods:
-            prompt_parts.append(
-                "\n=== RELEVANT METHODS/FUNCTIONS ===\n"
-                + "\n".join(f"- {method}" for method in relevant_methods)
-                + "\n=== END METHODS ==="
+            next_steps = (
+                f"PAUSE INVESTIGATION. Before calling {self.get_name()} step {step_number + 1}, you MUST examine code. "
+                + "Required: " + ", ".join(required_actions[:2]) + ". "
+                + f"Your next {self.get_name()} call (step_number: {step_number + 1}) must include "
+                f"NEW evidence from actual code examination, not just theories. If no bug evidence "
+                f"is found, suggesting "
+                f"collaboration with thought partner is valuable. NO recursive {self.get_name()} calls "
+                f"without investigation work!"
             )
 
-        if final_hypothesis:
-            prompt_parts.append(f"\n=== FINAL HYPOTHESIS ===\n{final_hypothesis}\n=== END HYPOTHESIS ===")
+        return {"next_steps": next_steps}
 
-        if images:
-            prompt_parts.append(
-                "\n=== VISUAL DEBUGGING INFORMATION ===\n"
-                + "\n".join(f"- {img}" for img in images)
-                + "\n=== END VISUAL INFORMATION ==="
-            )
 
-        # Add file content if we have relevant files
-        if relevant_files:
-            file_content, _ = self._prepare_file_content_for_prompt(relevant_files, None, "Essential debugging files")
-            if file_content:
-                prompt_parts.append(
-                    f"\n=== ESSENTIAL FILES FOR DEBUGGING ===\n{file_content}\n=== END ESSENTIAL FILES ==="
-                )
 
-        full_prompt = "\n".join(prompt_parts)
+    # Hook method overrides for debug-specific behavior
 
-        # Generate AI response
+    def prepare_step_data(self, request) -> dict:
+        """
+        Map debug-specific fields: relevant_methods -> relevant_context for internal processing.
+        """
+        step_data = {
+            "step": request.step,
+            "step_number": request.step_number,
+            "findings": request.findings,
+            "files_checked": request.files_checked,
+            "relevant_files": request.relevant_files,
+            "relevant_context": request.relevant_context,
+            "issues_found": [],  # Debug tool doesn't use issues_found field
+            "confidence": request.confidence,
+            "hypothesis": request.hypothesis,
+            "images": request.images or [],
+        }
+        return step_data
+
+    def should_skip_expert_analysis(self, request, consolidated_findings) -> bool:
+        """
+        Debug tool skips expert analysis when Claude has "certain" confidence.
+        """
+        return (
+            request.confidence == "certain" and
+            not request.next_step_required
+        )
+
+    # Override inheritance hooks for debug-specific behavior
+
+    def get_completion_status(self) -> str:
+        """Debug tools use debug-specific status."""
+        return "certain_confidence_proceed_with_fix"
+
+    def get_completion_data_key(self) -> str:
+        """Debug uses 'complete_investigation' key."""
+        return "complete_investigation"
+
+    def get_final_analysis_from_request(self, request):
+        """Debug tools use 'hypothesis' field."""
+        return request.hypothesis
+
+    def get_confidence_level(self, request) -> str:
+        """Debug tools use 'certain' for high confidence."""
+        return "certain"
+
+    def get_completion_message(self) -> str:
+        """Debug-specific completion message."""
+        return (
+            "Investigation complete with CERTAIN confidence. You have identified the exact "
+            "root cause and a minimal fix. MANDATORY: Present the user with the root cause analysis "
+            "and IMMEDIATELY proceed with implementing the simple fix without requiring further "
+            "consultation. Focus on the precise, minimal change needed."
+        )
+
+    def get_skip_reason(self) -> str:
+        """Debug-specific skip reason."""
+        return "Claude identified exact root cause with minimal fix requirement"
+
+    def get_request_relevant_context(self, request) -> list:
+        """Get relevant_context for debug tool."""
         try:
-            full_analysis_prompt = f"{self.get_system_prompt()}\n\n{full_prompt}\n\nPlease debug this issue following the structured format in the system prompt."
+            return request.relevant_context or []
+        except AttributeError:
+            return []
 
-            # Prepare generation kwargs
-            generation_kwargs = {
-                "prompt": full_analysis_prompt,
-                "model_name": model_name,
-                "system_prompt": "",  # Already included in prompt
-                "temperature": self.get_default_temperature(),
-                "thinking_mode": "high",  # High thinking for debug analysis
-            }
+    def get_skip_expert_analysis_status(self) -> str:
+        """Debug-specific expert analysis skip status."""
+        return "skipped_due_to_certain_confidence"
 
-            # Add images if available
-            if images:
-                generation_kwargs["images"] = images
+    def prepare_work_summary(self) -> str:
+        """Debug-specific work summary."""
+        return self._build_investigation_summary(self.consolidated_findings)
 
-            model_response = provider.generate_content(**generation_kwargs)
+    def get_completion_next_steps_message(self) -> str:
+        """
+        Debug-specific completion message.
+        """
+        return (
+            "INVESTIGATION IS COMPLETE. YOU MUST now summarize and present ALL key findings, confirmed "
+            "hypotheses, and exact recommended fixes. Clearly identify the most likely root cause and "
+            "provide concrete, actionable implementation guidance. Highlight affected code paths and display "
+            "reasoning that led to this conclusion—make it easy for a developer to understand exactly where "
+            "the problem lies."
+        )
 
-            if model_response.content:
-                # Try to parse as JSON
-                try:
-                    analysis_result = json.loads(model_response.content.strip())
-                    return analysis_result
-                except json.JSONDecodeError:
-                    # Return as text if not valid JSON
-                    return {
-                        "status": "analysis_complete",
-                        "raw_analysis": model_response.content,
-                        "parse_error": "Response was not valid JSON",
-                    }
-            else:
-                return {"error": "No response from model", "status": "empty_response"}
+    def get_step_guidance_message(self, request) -> str:
+        """
+        Debug-specific step guidance with detailed investigation instructions.
+        """
+        step_guidance = self.get_step_guidance(request.step_number, request.confidence, request)
+        return step_guidance["next_steps"]
 
-        except Exception as e:
-            logger.error(f"Error calling expert analysis: {e}", exc_info=True)
-            return {"error": str(e), "status": "analysis_error"}
+    def customize_workflow_response(self, response_data: dict, request) -> dict:
+        """
+        Customize response to match original debug tool format.
+        """
+        # Store initial issue on first step
+        if request.step_number == 1:
+            self.initial_issue = request.step
 
-    # Stub implementations for base class requirements
+        # Convert generic status names to debug-specific ones
+        tool_name = self.get_name()
+        status_mapping = {
+            f"{tool_name}_in_progress": "investigation_in_progress",
+            f"pause_for_{tool_name}": "pause_for_investigation",
+            f"{tool_name}_required": "investigation_required",
+            f"{tool_name}_complete": "investigation_complete",
+        }
+
+        if response_data["status"] in status_mapping:
+            response_data["status"] = status_mapping[response_data["status"]]
+
+        # Rename status field to match debug tool
+        if f"{tool_name}_status" in response_data:
+            response_data["investigation_status"] = response_data.pop(f"{tool_name}_status")
+            # Map relevant_context back to relevant_methods in status
+            if "relevant_context" in response_data["investigation_status"]:
+                response_data["investigation_status"]["relevant_methods"] = response_data[
+                    "investigation_status"
+                ].pop("relevant_context")
+                # Add debug-specific status fields
+                response_data["investigation_status"]["hypotheses_formed"] = len(self.consolidated_findings.hypotheses)
+
+        # Map relevant_context back to relevant_methods in complete investigation
+        if f"complete_{tool_name}" in response_data:
+            response_data["complete_investigation"] = response_data.pop(f"complete_{tool_name}")
+            if "relevant_context" in response_data["complete_investigation"]:
+                response_data["complete_investigation"]["relevant_methods"] = response_data[
+                    "complete_investigation"
+                ].pop("relevant_context")
+
+        # Map the completion flag to match original debug tool
+        if f"{tool_name}_complete" in response_data:
+            response_data["investigation_complete"] = response_data.pop(f"{tool_name}_complete")
+
+        return response_data
+
+    # Required abstract methods from BaseTool
+    def get_request_model(self):
+        """Return the debug-specific request model."""
+        return DebugInvestigationRequest
+
     async def prepare_prompt(self, request) -> str:
-        return ""  # Not used - execute() is overridden
-
-    def format_response(self, response: str, request, model_info: dict = None) -> str:
-        return response  # Not used - execute() is overridden
+        """Not used - workflow tools use execute_workflow()."""
+        return ""  # Workflow tools use execute_workflow() directly
